refactor: move cardimage component to global component folder
